+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags.Tests/Fakes/BrokenFakeResource.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace Phnx.AspNetCore.ETags.Tests.Fakes
+{
+    public class BrokenFakeResource
+    {
+        [ConcurrencyCheck]
+        public string Token => throw new NotImplementedException();
+    }
+}
