+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags.Tests/Fakes/FakeDto.cs
@@ -1,0 +1,16 @@
+﻿namespace Phnx.AspNetCore.ETags.Tests.Fakes
+{
+    public class FakeDto
+    {
+        public FakeDto()
+        {
+        }
+
+        public FakeDto(FakeResource resource)
+        {
+            Id = resource.Id;
+        }
+
+        public int Id { get; set; }
+    }
+}
