+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags.Tests/Fakes/FakeHeaderDictionary.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Primitives;
+using System;
+using System.Collections.Generic;
+
+namespace Phnx.AspNetCore.ETags.Tests.Fakes
+{
+    public class FakeHeaderDictionary : Dictionary<string, StringValues>, IHeaderDictionary
+    {
+        public long? ContentLength
+        {
+            get => throw new NotImplementedException();
+            set => throw new NotImplementedException();
+        }
+    }
+}
