+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags.Tests/Fakes/FakeResourceMap.cs
@@ -1,0 +1,25 @@
+﻿using Phnx.AspNetCore.ETags.Services;
+
+namespace Phnx.AspNetCore.ETags.Tests.Fakes
+{
+    public class FakeResourceMap : IResourceMap<FakeResource, FakeDto, FakePatch>
+    {
+        public FakeResource MapToData(FakeDto dto)
+        {
+            if (dto is null) return null;
+
+            return new FakeResource(dto.Id);
+        }
+
+        public FakeDto MapToDto(FakeResource data)
+        {
+            if (data is null) return null;
+
+            return new FakeDto(data);
+        }
+
+        public void PatchToData(FakePatch patch, FakeResource data)
+        {
+        }
+    }
+}
