+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Phnx.AspNetCore.ETags", "Phnx.AspNetCore.ETags\Phnx.AspNetCore.ETags.csproj", "{33EF0E41-7360-4CBA-AC6F-09A0E845E837}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Phnx.AspNetCore.ETags.Tests", "Phnx.AspNetCore.ETags.Tests\Phnx.AspNetCore.ETags.Tests.csproj", "{14D96AA1-CBDD-4EF5-A766-ED6D51B746AB}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{33EF0E41-7360-4CBA-AC6F-09A0E845E837}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{33EF0E41-7360-4CBA-AC6F-09A0E845E837}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{33EF0E41-7360-4CBA-AC6F-09A0E845E837}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{33EF0E41-7360-4CBA-AC6F-09A0E845E837}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14D96AA1-CBDD-4EF5-A766-ED6D51B746AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14D96AA1-CBDD-4EF5-A766-ED6D51B746AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14D96AA1-CBDD-4EF5-A766-ED6D51B746AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14D96AA1-CBDD-4EF5-A766-ED6D51B746AB}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {17C6FD8B-3609-4658-804E-62BA763159C1}
+	EndGlobalSection
+EndGlobal
