+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags/Models/ResourceCreatedResult.cs
