+++ src/AspNetCore/ETags/Phnx.AspNetCore.ETags/Phnx.AspNetCore.ETags.csproj
