--- src/AspNetCore/Rest/Phnx.AspNetCore.Rest.Tests/Fakes/BrokenFakeResource.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace Phnx.AspNetCore.Rest.Tests.Fakes
-{
-    public class BrokenFakeResource
-    {
-        [ConcurrencyCheck]
-        public string Token => throw new NotImplementedException();
-    }
-}
