--- src/AspNetCore/Rest/Phnx.AspNetCore.Rest.Tests/Fakes/FakeDto.cs
@@ -1,16 +1,0 @@
-﻿namespace Phnx.AspNetCore.Rest.Tests.Fakes
-{
-    public class FakeDto
-    {
-        public FakeDto()
-        {
-        }
-
-        public FakeDto(FakeResource resource)
-        {
-            Id = resource.Id;
-        }
-
-        public int Id { get; set; }
-    }
-}
