--- src/AspNetCore/Rest/Phnx.AspNetCore.Rest.Tests/Fakes/FakeHeaderDictionary.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Primitives;
-using System;
-using System.Collections.Generic;
-
-namespace Phnx.AspNetCore.Rest.Tests.Fakes
-{
-    public class FakeHeaderDictionary : Dictionary<string, StringValues>, IHeaderDictionary
-    {
-        public long? ContentLength
-        {
-            get => throw new NotImplementedException();
-            set => throw new NotImplementedException();
-        }
-    }
-}
