--- src/AspNetCore/Rest/Phnx.AspNetCore.Rest.Tests/Fakes/FakeResourcePatch.cs
@@ -1,16 +1,0 @@
-﻿namespace Phnx.AspNetCore.Rest.Tests.Fakes
-{
-    public class FakePatch
-    {
-        public FakePatch()
-        {
-        }
-
-        public FakePatch(FakeResource resource)
-        {
-            Id = resource.Id;
-        }
-
-        public int Id { get; set; }
-    }
-}
