--- src/AspNetCore/Rest/Phnx.AspNetCore.Rest.Tests/Phnx.AspNetCore.Rest.Tests.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Moq" Version="4.10.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
-    <PackageReference Include="nunit" Version="3.11.0" />
-    <PackageReference Include="NUnit3TestAdapter" Version="3.11.2" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Phnx.AspNetCore.Rest\Phnx.AspNetCore.Rest.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Phnx.Serialization">
-      <HintPath>..\..\..\Serialization\Phnx.Serialization\bin\Debug\netstandard2.0\Phnx.Serialization.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-</Project>
